Implement geotag logic

+++ pet_management/src/restapi/dto/community/PostCreateDto.swift
@@ -10,8 +10,8 @@ struct PostCreateParam: Encodable {
     var contents: String = "";
     var hashTags: [String] = [];
     var disclosure: String = "PUBLIC";
-    var geoTagLat: Float = 0;
-    var geoTagLong: Float = 0;
+    var geoTagLat: Double = 0;
+    var geoTagLong: Double = 0;
 }
 
 struct PostCreateDto: Decodable {
